# internal/api/openapi.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
)

const (
	ApiKeyScopes = "apiKey.Scopes"
)

// Defines values for CheckRequestHandles.
const (
	CheckRequestHandlesGithub CheckRequestHandles = "github"
)

// Defines values for CheckRequestProfile.
const (
	CheckRequestProfileDeveloper CheckRequestProfile = "developer"
	CheckRequestProfileMinimal   CheckRequestProfile = "minimal"
	CheckRequestProfileStartup   CheckRequestProfile = "startup"
	CheckRequestProfileWeb3      CheckRequestProfile = "web3"
	CheckRequestProfileWebsite   CheckRequestProfile = "website"
)

// Defines values for CheckRequestRegistries.
const (
	CheckRequestRegistriesCargo CheckRequestRegistries = "cargo"
	CheckRequestRegistriesNpm   CheckRequestRegistries = "npm"
	CheckRequestRegistriesPypi  CheckRequestRegistries = "pypi"
)

// Defines values for CheckResultAvailable.
const (
	CheckResultAvailableAvailable   CheckResultAvailable = "available"
	CheckResultAvailableError       CheckResultAvailable = "error"
	CheckResultAvailableRateLimited CheckResultAvailable = "rate_limited"
	CheckResultAvailableTaken       CheckResultAvailable = "taken"
	CheckResultAvailableUnknown     CheckResultAvailable = "unknown"
	CheckResultAvailableUnsupported CheckResultAvailable = "unsupported"
)

// Defines values for CheckResultCheckType.
const (
	CheckResultCheckTypeCargo  CheckResultCheckType = "cargo"
	CheckResultCheckTypeDomain CheckResultCheckType = "domain"
	CheckResultCheckTypeGithub CheckResultCheckType = "github"
	CheckResultCheckTypeNpm    CheckResultCheckType = "npm"
	CheckResultCheckTypePypi   CheckResultCheckType = "pypi"
)

// Defines values for CheckSummaryRiskLevel.
const (
	CheckSummaryRiskLevelHigh   CheckSummaryRiskLevel = "high"
	CheckSummaryRiskLevelLow    CheckSummaryRiskLevel = "low"
	CheckSummaryRiskLevelMedium CheckSummaryRiskLevel = "medium"
)

// Defines values for CompareRequestHandles.
const (
	Github CompareRequestHandles = "github"
)

// Defines values for CompareRequestProfile.
const (
	CompareRequestProfileDeveloper CompareRequestProfile = "developer"
	CompareRequestProfileMinimal   CompareRequestProfile = "minimal"
	CompareRequestProfileStartup   CompareRequestProfile = "startup"
	CompareRequestProfileWeb3      CompareRequestProfile = "web3"
	CompareRequestProfileWebsite   CompareRequestProfile = "website"
)

// Defines values for CompareRequestRegistries.
const (
	CompareRequestRegistriesCargo CompareRequestRegistries = "cargo"
	CompareRequestRegistriesNpm   CompareRequestRegistries = "npm"
	CompareRequestRegistriesPypi  CompareRequestRegistries = "pypi"
)

// Defines values for ExpertAnalysisRiskLevel.
const (
	ExpertAnalysisRiskLevelHigh   ExpertAnalysisRiskLevel = "high"
	ExpertAnalysisRiskLevelLow    ExpertAnalysisRiskLevel = "low"
	ExpertAnalysisRiskLevelMedium ExpertAnalysisRiskLevel = "medium"
)

// Defines values for HealthCheckStatus.
const (
	Fail HealthCheckStatus = "fail"
	Pass HealthCheckStatus = "pass"
	Warn HealthCheckStatus = "warn"
)

// Defines values for HealthResponseStatus.
const (
	Degraded  HealthResponseStatus = "degraded"
	Healthy   HealthResponseStatus = "healthy"
	Unhealthy HealthResponseStatus = "unhealthy"
)

// CheckRequest defines model for CheckRequest.
type CheckRequest struct {
	// Expert Include AI-powered brand safety analysis
	Expert *bool `json:"expert,omitempty"`

	// Handles Social handles to check (overrides profile)
	Handles *[]CheckRequestHandles `json:"handles,omitempty"`

	// Name The name to check
	Name string `json:"name"`

	// Profile Predefined check profile to use
	Profile *CheckRequestProfile `json:"profile,omitempty"`

	// Registries Package registries to check (overrides profile)
	Registries *[]CheckRequestRegistries `json:"registries,omitempty"`

	// Tlds Custom TLDs to check (overrides profile)
	Tlds *[]string `json:"tlds,omitempty"`
}

// CheckRequestHandles defines model for CheckRequest.Handles.
type CheckRequestHandles string

// CheckRequestProfile Predefined check profile to use
type CheckRequestProfile string

// CheckRequestRegistries defines model for CheckRequest.Registries.
type CheckRequestRegistries string

// CheckResponse defines model for CheckResponse.
type CheckResponse struct {
	Expert *ExpertAnalysis `json:"expert,omitempty"`

	// Name The name that was checked
	Name string `json:"name"`

	// Results Individual check results
	Results []CheckResult `json:"results"`
	Summary CheckSummary  `json:"summary"`
}

// CheckResult defines model for CheckResult.
type CheckResult struct {
	// Available Availability status
	Available CheckResultAvailable `json:"available"`

	// CheckType Type of check performed
	CheckType CheckResultCheckType `json:"check_type"`

	// Message Additional information or error message
	Message *string `json:"message,omitempty"`

	// Name Full name checked (e.g., acmecorp.com)
	Name       string      `json:"name"`
	Provenance *Provenance `json:"provenance,omitempty"`

	// Tld TLD for domain checks
	Tld *string `json:"tld,omitempty"`
}

// CheckResultAvailable Availability status
type CheckResultAvailable string

// CheckResultCheckType Type of check performed
type CheckResultCheckType string

// CheckSummary defines model for CheckSummary.
type CheckSummary struct {
	// Available Number of available results
	Available int `json:"available"`

	// RiskLevel Overall risk assessment
	RiskLevel *CheckSummaryRiskLevel `json:"risk_level,omitempty"`

	// Taken Number of taken results
	Taken int `json:"taken"`

	// Total Total checks performed
	Total int `json:"total"`

	// Unknown Number of unknown/error results
	Unknown int `json:"unknown"`
}

// CheckSummaryRiskLevel Overall risk assessment
type CheckSummaryRiskLevel string

// CompareCandidate defines model for CompareCandidate.
type CompareCandidate struct {
	Expert  *ExpertAnalysis `json:"expert,omitempty"`
	Name    string          `json:"name"`
	Results []CheckResult   `json:"results"`
	Summary CheckSummary    `json:"summary"`
}

// CompareRequest defines model for CompareRequest.
type CompareRequest struct {
	// Expert Include AI analysis for each candidate
	Expert  *bool                    `json:"expert,omitempty"`
	Handles *[]CompareRequestHandles `json:"handles,omitempty"`

	// Names Names to compare
	Names []string `json:"names"`

	// Profile Check profile to use
	Profile    *CompareRequestProfile      `json:"profile,omitempty"`
	Registries *[]CompareRequestRegistries `json:"registries,omitempty"`

	// Tlds Custom TLDs (overrides profile)
	Tlds *[]string `json:"tlds,omitempty"`
}

// CompareRequestHandles defines model for CompareRequest.Handles.
type CompareRequestHandles string

// CompareRequestProfile Check profile to use
type CompareRequestProfile string

// CompareRequestRegistries defines model for CompareRequest.Registries.
type CompareRequestRegistries string

// CompareResponse defines model for CompareResponse.
type CompareResponse struct {
	Candidates []CompareCandidate `json:"candidates"`

	// Recommendation AI recommendation (if expert=true)
	Recommendation *string `json:"recommendation,omitempty"`
}

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code string `json:"code"`

	// Details Additional error details
	Details *map[string]interface{} `json:"details,omitempty"`

	// Message Human-readable error message
	Message string `json:"message"`

	// RetryAfter Seconds to wait before retrying (for rate limits)
	RetryAfter *int `json:"retry_after,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error Error `json:"error"`
}

// ExpertAnalysis defines model for ExpertAnalysis.
type ExpertAnalysis struct {
	// Details Detailed analysis
	Details *string `json:"details,omitempty"`

	// Provider AI provider used (e.g., xai, openai)
	Provider *string `json:"provider,omitempty"`

	// RiskLevel AI-assessed risk level
	RiskLevel *ExpertAnalysisRiskLevel `json:"risk_level,omitempty"`

	// Summary Brief risk summary
	Summary *string `json:"summary,omitempty"`
}

// ExpertAnalysisRiskLevel AI-assessed risk level
type ExpertAnalysisRiskLevel string

// HealthCheck defines model for HealthCheck.
type HealthCheck struct {
	Message *string           `json:"message,omitempty"`
	Status  HealthCheckStatus `json:"status"`
}

// HealthCheckStatus defines model for HealthCheck.Status.
type HealthCheckStatus string

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	// Checks Individual component health checks
	Checks *map[string]HealthCheck `json:"checks,omitempty"`

	// Status Overall health status
	Status HealthResponseStatus `json:"status"`

	// Version Server version
	Version *string `json:"version,omitempty"`
}

// HealthResponseStatus Overall health status
type HealthResponseStatus string

// Profile defines model for Profile.
type Profile struct {
	// Description Human-readable description
	Description *string `json:"description,omitempty"`

	// Handles Social handles this profile checks
	Handles *[]string `json:"handles,omitempty"`

	// IsBuiltin Whether this is a built-in profile
	IsBuiltin *bool `json:"is_builtin,omitempty"`

	// Name Profile identifier
	Name string `json:"name"`

	// Registries Package registries this profile checks
	Registries *[]string `json:"registries,omitempty"`

	// Tlds TLDs this profile checks
	Tlds *[]string `json:"tlds,omitempty"`
}

// ProfileListResponse defines model for ProfileListResponse.
type ProfileListResponse struct {
	Profiles []Profile `json:"profiles"`
}

// Provenance defines model for Provenance.
type Provenance struct {
	CacheExpiresAt *time.Time `json:"cache_expires_at,omitempty"`

	// FromCache Whether result came from cache
	FromCache   *bool      `json:"from_cache,omitempty"`
	RequestedAt *time.Time `json:"requested_at,omitempty"`
	ResolvedAt  *time.Time `json:"resolved_at,omitempty"`

	// Server Server that provided the result
	Server *string `json:"server,omitempty"`

	// Source How the result was determined
	Source *string `json:"source,omitempty"`
}

// ProviderStatus defines model for ProviderStatus.
type ProviderStatus struct {
	// Available Whether the provider is available
	Available bool `json:"available"`

	// Message Status message or error
	Message *string `json:"message,omitempty"`
}

// RateLimitStatus defines model for RateLimitStatus.
type RateLimitStatus struct {
	// Remaining Remaining requests in current window
	Remaining *int `json:"remaining,omitempty"`

	// ResetAt When the rate limit window resets
	ResetAt *time.Time `json:"reset_at,omitempty"`
}

// StatusResponse defines model for StatusResponse.
type StatusResponse struct {
	// Providers Status of each check provider
	Providers map[string]ProviderStatus `json:"providers"`

	// RateLimits Current rate limit headroom
	RateLimits *map[string]RateLimitStatus `json:"rate_limits,omitempty"`
}

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// RateLimited defines model for RateLimited.
type RateLimited = ErrorResponse

// Unauthorized defines model for Unauthorized.
type Unauthorized = ErrorResponse

// CheckNameJSONRequestBody defines body for CheckName for application/json ContentType.
type CheckNameJSONRequestBody = CheckRequest

// CompareCandidatesJSONRequestBody defines body for CompareCandidates for application/json ContentType.
type CompareCandidatesJSONRequestBody = CompareRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Health check
	// (GET /health)
	GetHealth(w http.ResponseWriter, r *http.Request)
	// Check name availability
	// (POST /v1/check)
	CheckName(w http.ResponseWriter, r *http.Request)
	// Compare multiple name candidates
	// (POST /v1/compare)
	CompareCandidates(w http.ResponseWriter, r *http.Request)
	// List available profiles
	// (GET /v1/profiles)
	ListProfiles(w http.ResponseWriter, r *http.Request)
	// Get server status
	// (GET /v1/status)
	GetStatus(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Health check
// (GET /health)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Check name availability
// (POST /v1/check)
func (_ Unimplemented) CheckName(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Compare multiple name candidates
// (POST /v1/compare)
func (_ Unimplemented) CompareCandidates(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List available profiles
// (GET /v1/profiles)
func (_ Unimplemented) ListProfiles(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get server status
// (GET /v1/status)
func (_ Unimplemented) GetStatus(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CheckName operation middleware
func (siw *ServerInterfaceWrapper) CheckName(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CheckName(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CompareCandidates operation middleware
func (siw *ServerInterfaceWrapper) CompareCandidates(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CompareCandidates(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListProfiles operation middleware
func (siw *ServerInterfaceWrapper) ListProfiles(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListProfiles(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetStatus operation middleware
func (siw *ServerInterfaceWrapper) GetStatus(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, ApiKeyScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetStatus(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.GetHealth)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/check", wrapper.CheckName)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/compare", wrapper.CompareCandidates)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/profiles", wrapper.ListProfiles)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/status", wrapper.GetStatus)
	})

	return r
}